arch/aarch64: Support arguments and return values (WIP)
This is only a partial support in that the first argument will not be
correct and floating-point arguments are not supported yet.  The first
argument is clobbered by the compiler before entering to the "_mcount"
function (it will have the address of parent function).

Also I cannot access the FP registers using inline assembly like below.
I think it's just a simple issue but I couldn't find a way to fix it.

    CC FPIC   arch/aarch64/mcount-support.op
  /tmp/cc9kumMH.s: Assembler messages:
  /tmp/cc9kumMH.s:420: Error: operand 1 must be an integer register -- `str %s0,[x0]'
  /tmp/cc9kumMH.s:422: Error: operand 1 must be an integer register -- `str %d0,[x0]'

Signed-off-by: Namhyung Kim <namhyung@gmail.com>